requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyascii==0.3.2
apispec==6.4.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contractions==0.1.73
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
filelock==3.13.1
Flask==3.0.2
flask-apispec==0.11.4
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
fsspec==2024.2.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.20.3
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
jellyfish==1.0.3
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-enum==1.5.1
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pickleDB==0.9.2
pillow==10.2.0
preshed==3.0.9
pyahocorasick==2.0.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pyparsing==3.1.1
pyspellchecker==0.8.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rdflib==7.0.0
regex==2023.12.25
requests==2.31.0
runstats==2.0.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
segtok==1.5.11
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
socketIO-client==0.7.2
soupsieve==2.5
spacy==3.7.4
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SPARQLWrapper==2.0.0
srsly==2.4.8
sympy==1.12
tabulate==0.9.0
textsearch==0.0.24
thinc==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.1
typer==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
webargs==8.4.0
websocket-client==1.7.0
Werkzeug==3.0.1
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yake==0.4.8
zope.event==5.0
zope.interface==6.2